Found and Implemented the Filter Bank method of preprocessing and a Filter Bank CSP method. The source for this is: https://github.com/NeuroTechX/moabb/blob/be1f81220869158ef37e1ab91b0279fe60aeed5b/pipelines/FBCSP.py
Implemented Filter Bank process into the preprocessing of the ClassifierTester class. Split this into multiple sub-processes to aid in readability and cut down the size of __init__.
Still need to put in the actual filter bank CSP into ClassifierTester, along with checks to make sure that the filter bank processed data is only used on filter bank pipelines, not the normal ones.